Reinitialize forecast parameters

+++ src/lib/models/losses.py
@@ -145,7 +145,7 @@ class RegL1Loss(nn.Module):
     mask = mask.unsqueeze(2).expand_as(pred).float()
     # loss = F.l1_loss(pred * mask, target * mask, reduction='elementwise_mean')
     loss = F.l1_loss(pred * mask, target * mask, size_average=False)
-    loss = loss / (mask.sum() + 1e-4)
+    loss = loss / (mask.sum().float() + 1e-4)
     return loss
 
 class NormRegL1Loss(nn.Module):
